backend/python-easy-ocr/requirements.txt
easyocr
flask
pillow
numpy
passporteye
